name changes on propagation files
